JavaScript bootstrap
This heavily depends on the internals of the current Forth VM. It was
tested with hsoft/collapseos@723d3c424987035be280a13f65f2d2222c7d643f,
and may not work with other commits. Note that I may not update this in
the future.

It can be used to bootstrap the boot binary for the JavaScript VM
(emul0-xcomp.fs) without depending on any opaque binaries (like the ones
from the CollapseOS repository) and it will create a boot binary that is
100% identical to the one built by the "official" stage compiler.

I think it is quite remarkable that the Forth VM is <250 lines JS and
the bootstrap.js is <500 lines.