Add support for gzip DecompressionStream
When loading the provided ROM or SD card images, if the browser supports gzip
DecompressionStream, download it as .rom.gz or .bfs.gz instead of unwrapping a
.png file. Smaller file size, faster to decompress, and Firefox's GPU
acceleration will not mess with the pixels, which would result in boot failure.

In browsers that don't support it, fall back to the existing code path.